+++ config/autoload/redis.global.php
@@ -6,9 +6,14 @@ use Shlinkio\Shlink\Core\Config\EnvVars;
 
 return (static function (): array {
     $redisServers = EnvVars::REDIS_SERVERS->loadFromEnv();
+    $pubSub = [
+        'redis' => [
+            'pub_sub_enabled' => $redisServers !== null && EnvVars::REDIS_PUB_SUB_ENABLED->loadFromEnv(false),
+        ],
+    ];
 
     return match ($redisServers) {
-        null => [],
+        null => $pubSub,
         default => [
             'cache' => [
                 'redis' => [
@@ -16,6 +21,7 @@ return (static function (): array {
                     'sentinel_service' => EnvVars::REDIS_SENTINEL_SERVICE->loadFromEnv(),
                 ],
             ],
+            ...$pubSub,
         ],
     };
 })();