+++ module/Core/config/event_dispatcher.config.php
@@ -7,10 +7,11 @@ namespace Shlinkio\Shlink\Core;
 use Laminas\ServiceManager\AbstractFactory\ConfigAbstractFactory;
 use Psr\EventDispatcher\EventDispatcherInterface;
 use Shlinkio\Shlink\CLI\Util\GeolocationDbUpdater;
+use Shlinkio\Shlink\Common\Cache\RedisPublishingHelper;
+use Shlinkio\Shlink\Common\Mercure\MercureHubPublishingHelper;
 use Shlinkio\Shlink\Common\RabbitMq\RabbitMqPublishingHelper;
 use Shlinkio\Shlink\IpGeolocation\GeoLite2\DbUpdater;
 use Shlinkio\Shlink\IpGeolocation\Resolver\IpLocationResolverInterface;
-use Symfony\Component\Mercure\Hub;
 
 return [
 
@@ -24,12 +25,14 @@ return [
             EventDispatcher\Event\VisitLocated::class => [
                 EventDispatcher\Mercure\NotifyVisitToMercure::class,
                 EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class,
+                EventDispatcher\RedisPubSub\NotifyVisitToRedis::class,
                 EventDispatcher\NotifyVisitToWebHooks::class,
                 EventDispatcher\UpdateGeoLiteDb::class,
             ],
             EventDispatcher\Event\ShortUrlCreated::class => [
                 EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class,
                 EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class,
+                EventDispatcher\RedisPubSub\NotifyNewShortUrlToRedis::class,
             ],
         ],
     ],
@@ -42,6 +45,8 @@ return [
             EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class => ConfigAbstractFactory::class,
             EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class => ConfigAbstractFactory::class,
             EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => ConfigAbstractFactory::class,
+            EventDispatcher\RedisPubSub\NotifyVisitToRedis::class => ConfigAbstractFactory::class,
+            EventDispatcher\RedisPubSub\NotifyNewShortUrlToRedis::class => ConfigAbstractFactory::class,
             EventDispatcher\UpdateGeoLiteDb::class => ConfigAbstractFactory::class,
         ],
 
@@ -56,6 +61,12 @@ return [
                 EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
             ],
             EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => [
+                EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
+            ],
+            EventDispatcher\RedisPubSub\NotifyVisitToRedis::class => [
+                EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
+            ],
+            EventDispatcher\RedisPubSub\NotifyNewShortUrlToRedis::class => [
                 EventDispatcher\CloseDbConnectionEventListenerDelegator::class,
             ],
             EventDispatcher\NotifyVisitToWebHooks::class => [
@@ -81,31 +92,45 @@ return [
             Options\AppOptions::class,
         ],
         EventDispatcher\Mercure\NotifyVisitToMercure::class => [
-            Hub::class,
-            Mercure\MercureUpdatesGenerator::class,
+            MercureHubPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
             'em',
             'Logger_Shlink',
         ],
         EventDispatcher\Mercure\NotifyNewShortUrlToMercure::class => [
-            Hub::class,
-            Mercure\MercureUpdatesGenerator::class,
+            MercureHubPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
             'em',
             'Logger_Shlink',
         ],
         EventDispatcher\RabbitMq\NotifyVisitToRabbitMq::class => [
             RabbitMqPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
             'em',
             'Logger_Shlink',
             Visit\Transformer\OrphanVisitDataTransformer::class,
-            ShortUrl\Transformer\ShortUrlDataTransformer::class,
             Options\RabbitMqOptions::class,
         ],
         EventDispatcher\RabbitMq\NotifyNewShortUrlToRabbitMq::class => [
             RabbitMqPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
             'em',
             'Logger_Shlink',
-            ShortUrl\Transformer\ShortUrlDataTransformer::class,
             Options\RabbitMqOptions::class,
+        ],
+        EventDispatcher\RedisPubSub\NotifyVisitToRedis::class => [
+            RedisPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
+            'em',
+            'Logger_Shlink',
+            'config.redis.pub_sub_enabled',
+        ],
+        EventDispatcher\RedisPubSub\NotifyNewShortUrlToRedis::class => [
+            RedisPublishingHelper::class,
+            EventDispatcher\PublishingUpdatesGenerator::class,
+            'em',
+            'Logger_Shlink',
+            'config.redis.pub_sub_enabled',
         ],
         EventDispatcher\UpdateGeoLiteDb::class => [GeolocationDbUpdater::class, 'Logger_Shlink'],
     ],