+++ module/Core/src/EventDispatcher/Mercure/NotifyVisitToMercure.php
@@ -4,61 +4,18 @@ declare(strict_types=1);
 
 namespace Shlinkio\Shlink\Core\EventDispatcher\Mercure;
 
-use Doctrine\ORM\EntityManagerInterface;
-use Psr\Log\LoggerInterface;
-use Shlinkio\Shlink\Core\Entity\Visit;
-use Shlinkio\Shlink\Core\EventDispatcher\Event\VisitLocated;
-use Shlinkio\Shlink\Core\Mercure\MercureUpdatesGeneratorInterface;
-use Symfony\Component\Mercure\HubInterface;
-use Symfony\Component\Mercure\Update;
-use Throwable;
+use Shlinkio\Shlink\Core\EventDispatcher\Async\AbstractNotifyVisitListener;
+use Shlinkio\Shlink\Core\EventDispatcher\Async\RemoteSystem;
 
-use function Functional\each;
-
-class NotifyVisitToMercure
+class NotifyVisitToMercure extends AbstractNotifyVisitListener
 {
-    public function __construct(
-        private readonly HubInterface $hub,
-        private readonly MercureUpdatesGeneratorInterface $updatesGenerator,
-        private readonly EntityManagerInterface $em,
-        private readonly LoggerInterface $logger,
-    ) {
+    protected function isEnabled(): bool
+    {
+        return true;
     }
 
-    public function __invoke(VisitLocated $shortUrlLocated): void
+    protected function getRemoteSystem(): RemoteSystem
     {
-        $visitId = $shortUrlLocated->visitId;
-
-        /** @var Visit|null $visit */
-        $visit = $this->em->find(Visit::class, $visitId);
-        if ($visit === null) {
-            $this->logger->warning('Tried to notify mercure for visit with id "{visitId}", but it does not exist.', [
-                'visitId' => $visitId,
-            ]);
-            return;
-        }
-
-        try {
-            each($this->determineUpdatesForVisit($visit), fn (Update $update) => $this->hub->publish($update));
-        } catch (Throwable $e) {
-            $this->logger->debug('Error while trying to notify mercure hub with new visit. {e}', [
-                'e' => $e,
-            ]);
-        }
-    }
-
-    /**
-     * @return Update[]
-     */
-    private function determineUpdatesForVisit(Visit $visit): array
-    {
-        if ($visit->isOrphan()) {
-            return [$this->updatesGenerator->newOrphanVisitUpdate($visit)];
-        }
-
-        return [
-            $this->updatesGenerator->newShortUrlVisitUpdate($visit),
-            $this->updatesGenerator->newVisitUpdate($visit),
-        ];
+        return RemoteSystem::MERCURE;
     }
 }