+++ module/Core/src/EventDispatcher/PublishingUpdatesGenerator.php
@@ -2,17 +2,14 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\Mercure;
+namespace Shlinkio\Shlink\Core\EventDispatcher;
 
 use Shlinkio\Shlink\Common\Rest\DataTransformerInterface;
+use Shlinkio\Shlink\Common\UpdatePublishing\Update;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\Entity\Visit;
-use Shlinkio\Shlink\Core\EventDispatcher\Topic;
-use Symfony\Component\Mercure\Update;
 
-use function Shlinkio\Shlink\Common\json_encode;
-
-final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
+final class PublishingUpdatesGenerator implements PublishingUpdatesGeneratorInterface
 {
     public function __construct(
         private readonly DataTransformerInterface $shortUrlTransformer,
@@ -22,17 +19,17 @@ final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
 
     public function newVisitUpdate(Visit $visit): Update
     {
-        return new Update(Topic::NEW_VISIT->value, json_encode([
+        return Update::forTopicAndPayload(Topic::NEW_VISIT->value, [
             'shortUrl' => $this->shortUrlTransformer->transform($visit->getShortUrl()),
-            'visit' => $visit,
-        ]));
+            'visit' => $visit->jsonSerialize(),
+        ]);
     }
 
     public function newOrphanVisitUpdate(Visit $visit): Update
     {
-        return new Update(Topic::NEW_ORPHAN_VISIT->value, json_encode([
+        return Update::forTopicAndPayload(Topic::NEW_ORPHAN_VISIT->value, [
             'visit' => $this->orphanVisitTransformer->transform($visit),
-        ]));
+        ]);
     }
 
     public function newShortUrlVisitUpdate(Visit $visit): Update
@@ -40,16 +37,16 @@ final class MercureUpdatesGenerator implements MercureUpdatesGeneratorInterface
         $shortUrl = $visit->getShortUrl();
         $topic = Topic::newShortUrlVisit($shortUrl?->getShortCode());
 
-        return new Update($topic, json_encode([
+        return Update::forTopicAndPayload($topic, [
             'shortUrl' => $this->shortUrlTransformer->transform($shortUrl),
-            'visit' => $visit,
-        ]));
+            'visit' => $visit->jsonSerialize(),
+        ]);
     }
 
     public function newShortUrlUpdate(ShortUrl $shortUrl): Update
     {
-        return new Update(Topic::NEW_SHORT_URL->value, json_encode([
+        return Update::forTopicAndPayload(Topic::NEW_SHORT_URL->value, [
             'shortUrl' => $this->shortUrlTransformer->transform($shortUrl),
-        ]));
+        ]);
     }
 }