+++ module/Core/src/EventDispatcher/RedisPubSub/NotifyVisitToRedis.php
@@ -2,35 +2,34 @@
 
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\Core\EventDispatcher\RabbitMq;
+namespace Shlinkio\Shlink\Core\EventDispatcher\RedisPubSub;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Psr\Log\LoggerInterface;
 use Shlinkio\Shlink\Common\UpdatePublishing\PublishingHelperInterface;
-use Shlinkio\Shlink\Core\EventDispatcher\Async\AbstractNotifyNewShortUrlListener;
+use Shlinkio\Shlink\Core\EventDispatcher\Async\AbstractNotifyVisitListener;
 use Shlinkio\Shlink\Core\EventDispatcher\Async\RemoteSystem;
 use Shlinkio\Shlink\Core\EventDispatcher\PublishingUpdatesGeneratorInterface;
-use Shlinkio\Shlink\Core\Options\RabbitMqOptions;
 
-class NotifyNewShortUrlToRabbitMq extends AbstractNotifyNewShortUrlListener
+class NotifyVisitToRedis extends AbstractNotifyVisitListener
 {
     public function __construct(
-        PublishingHelperInterface $rabbitMqHelper,
+        PublishingHelperInterface $redisHelper,
         PublishingUpdatesGeneratorInterface $updatesGenerator,
         EntityManagerInterface $em,
         LoggerInterface $logger,
-        private readonly RabbitMqOptions $options,
+        private readonly bool $enabled,
     ) {
-        parent::__construct($rabbitMqHelper, $updatesGenerator, $em, $logger);
+        parent::__construct($redisHelper, $updatesGenerator, $em, $logger);
     }
 
     protected function isEnabled(): bool
     {
-        return $this->options->isEnabled();
+        return $this->enabled;
     }
 
     protected function getRemoteSystem(): RemoteSystem
     {
-        return RemoteSystem::RABBIT_MQ;
+        return RemoteSystem::REDIS_PUB_SUB;
     }
 }