+++ module/Core/test/EventDispatcher/Mercure/NotifyNewShortUrlToMercureTest.php
@@ -11,32 +11,32 @@ use Prophecy\Argument;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Prophecy\Prophecy\ObjectProphecy;
 use Psr\Log\LoggerInterface;
+use Shlinkio\Shlink\Common\UpdatePublishing\PublishingHelperInterface;
+use Shlinkio\Shlink\Common\UpdatePublishing\Update;
 use Shlinkio\Shlink\Core\Entity\ShortUrl;
 use Shlinkio\Shlink\Core\EventDispatcher\Event\ShortUrlCreated;
 use Shlinkio\Shlink\Core\EventDispatcher\Mercure\NotifyNewShortUrlToMercure;
-use Shlinkio\Shlink\Core\Mercure\MercureUpdatesGeneratorInterface;
-use Symfony\Component\Mercure\HubInterface;
-use Symfony\Component\Mercure\Update;
+use Shlinkio\Shlink\Core\EventDispatcher\PublishingUpdatesGeneratorInterface;
 
 class NotifyNewShortUrlToMercureTest extends TestCase
 {
     use ProphecyTrait;
 
     private NotifyNewShortUrlToMercure $listener;
-    private ObjectProphecy $hub;
+    private ObjectProphecy $helper;
     private ObjectProphecy $updatesGenerator;
     private ObjectProphecy $em;
     private ObjectProphecy $logger;
 
     protected function setUp(): void
     {
-        $this->hub = $this->prophesize(HubInterface::class);
-        $this->updatesGenerator = $this->prophesize(MercureUpdatesGeneratorInterface::class);
+        $this->helper = $this->prophesize(PublishingHelperInterface::class);
+        $this->updatesGenerator = $this->prophesize(PublishingUpdatesGeneratorInterface::class);
         $this->em = $this->prophesize(EntityManagerInterface::class);
         $this->logger = $this->prophesize(LoggerInterface::class);
 
         $this->listener = new NotifyNewShortUrlToMercure(
-            $this->hub->reveal(),
+            $this->helper->reveal(),
             $this->updatesGenerator->reveal(),
             $this->em->reveal(),
             $this->logger->reveal(),
@@ -52,10 +52,10 @@ class NotifyNewShortUrlToMercureTest extends TestCase
 
         $find->shouldHaveBeenCalledOnce();
         $this->logger->warning(
-            'Tried to notify Mercure for new short URL with id "{shortUrlId}", but it does not exist.',
-            ['shortUrlId' => '123'],
+            'Tried to notify {name} for new short URL with id "{shortUrlId}", but it does not exist.',
+            ['shortUrlId' => '123', 'name' => 'Mercure'],
         )->shouldHaveBeenCalledOnce();
-        $this->hub->publish(Argument::cetera())->shouldNotHaveBeenCalled();
+        $this->helper->publishUpdate(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->updatesGenerator->newShortUrlUpdate(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->logger->debug(Argument::cetera())->shouldNotHaveBeenCalled();
     }
@@ -64,17 +64,16 @@ class NotifyNewShortUrlToMercureTest extends TestCase
     public function expectedNotificationIsPublished(): void
     {
         $shortUrl = ShortUrl::withLongUrl('');
-        $update = new Update([]);
+        $update = Update::forTopicAndPayload('', []);
 
         $find = $this->em->find(ShortUrl::class, '123')->willReturn($shortUrl);
         $newUpdate = $this->updatesGenerator->newShortUrlUpdate($shortUrl)->willReturn($update);
-        $publish = $this->hub->publish($update)->willReturn('');
 
         ($this->listener)(new ShortUrlCreated('123'));
 
         $find->shouldHaveBeenCalledOnce();
         $newUpdate->shouldHaveBeenCalledOnce();
-        $publish->shouldHaveBeenCalledOnce();
+        $this->helper->publishUpdate($update)->shouldHaveBeenCalledOnce();
         $this->logger->warning(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->logger->debug(Argument::cetera())->shouldNotHaveBeenCalled();
     }
@@ -83,12 +82,12 @@ class NotifyNewShortUrlToMercureTest extends TestCase
     public function messageIsPrintedIfPublishingFails(): void
     {
         $shortUrl = ShortUrl::withLongUrl('');
-        $update = new Update([]);
+        $update = Update::forTopicAndPayload('', []);
         $e = new Exception('Error');
 
         $find = $this->em->find(ShortUrl::class, '123')->willReturn($shortUrl);
         $newUpdate = $this->updatesGenerator->newShortUrlUpdate($shortUrl)->willReturn($update);
-        $publish = $this->hub->publish($update)->willThrow($e);
+        $publish = $this->helper->publishUpdate($update)->willThrow($e);
 
         ($this->listener)(new ShortUrlCreated('123'));
 
@@ -97,8 +96,8 @@ class NotifyNewShortUrlToMercureTest extends TestCase
         $publish->shouldHaveBeenCalledOnce();
         $this->logger->warning(Argument::cetera())->shouldNotHaveBeenCalled();
         $this->logger->debug(
-            'Error while trying to notify mercure hub with new short URL. {e}',
-            ['e' => $e],
+            'Error while trying to notify {name} with new short URL. {e}',
+            ['e' => $e, 'name' => 'Mercure'],
         )->shouldHaveBeenCalledOnce();
     }
 }